requirements.txt
absl-py==0.11.0
asgiref==3.3.1
astunparse==1.6.3
boto3==1.17.5
botocore==1.20.5
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
Django==3.1.4
django-cors-headers==3.6.0
django-storages==1.11.1
djangorestframework==3.12.2
flatbuffers==1.12
gast==0.3.3
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.3.0
imutils==0.5.3
jmespath==0.10.0
Keras-Preprocessing==1.1.2
Markdown==3.3.3
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
pandas==1.2.0
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytesseract==0.3.6
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
